refactor: 폴더 구조 및 네이밍 수정
